Support uuid-in-package: URLs in Subresource WebBundles
This allows subresource and subframe loading from WebBundles with
uuid-in-package: scheme URLs. We have been using urn:uuid: for this
purpose, but since urn: scheme can be handled by custom protocol
handler, there was a possibility of conflict.

Explainer change: https://github.com/WICG/webpackage/pull/677

The semantics is the same as currently we have for urn:uuid.

Temporarily, we support both uuid-in-package: and urn:uuid: resources so
that this will not break currently running origin trial. We will remove
the urn:uuid: support before shipping this feature.

Bug: 1257045
Change-Id: Ie86d3a3207fdaa6a4b0b4337cac2f1c6457e132b